refactor: GroupInvite moved to model directory
